make reachables accessable records
