+++ src/main/java/de/feu/massim22/group3/agents/Belief.java
@@ -8,9 +8,11 @@ import java.util.Map;
 import java.util.Set;
 import java.awt.Point;
 
+import de.feu.massim22.group3.agents.Reachable.ReachableDispenser;
+import de.feu.massim22.group3.agents.Reachable.ReachableGoalZone;
+import de.feu.massim22.group3.agents.Reachable.ReachableRoleZone;
 import de.feu.massim22.group3.map.CellType;
 import de.feu.massim22.group3.map.ZoneType;
-import de.feu.massim22.group3.agents.BeliefTypes.*;
 import de.feu.massim22.group3.utils.logging.AgentLogger;
 import eis.iilang.Function;
 import eis.iilang.Identifier;
@@ -613,5 +615,3 @@ public class Belief {
         }
     }*/
 }
-
-
