+++ src/main/java/de/feu/massim22/group3/agents/Desires/ADesires/DigFree.java
@@ -6,8 +6,8 @@ import java.util.List;
 import de.feu.massim22.group3.agents.BdiAgentV2;
 import de.feu.massim22.group3.agents.DesireUtilities;
 import de.feu.massim22.group3.agents.DirectionUtil;
-import de.feu.massim22.group3.agents.BeliefTypes.*;
-import de.feu.massim22.group3.agents.Belief.*;
+import de.feu.massim22.group3.agents.Reachable.ReachableDispenser;
+import de.feu.massim22.group3.agents.Reachable.ReachableGoalZone;
 import eis.iilang.Action;
 import eis.iilang.Identifier;
 import massim.protocol.data.Thing;
