+++ src/main/java/de/feu/massim22/group3/agents/Desires/SubDesires/SubDesire.java
@@ -1,8 +1,5 @@
 package de.feu.massim22.group3.agents.Desires.SubDesires;
 
-import java.util.LinkedList;
-
-import de.feu.massim22.group3.agents.Desires.Desires;
 import de.feu.massim22.group3.agents.BdiAgent;
 import eis.iilang.Action;
