+++ src/main/java/de/feu/massim22/group3/agents/Reachable/ReachableRoleZone.java
@@ -1,0 +1,12 @@
+package de.feu.massim22.group3.agents.Reachable;
+
+import de.feu.massim22.group3.agents.DirectionUtil;
+
+import java.awt.Point;
+public record ReachableRoleZone(Point position, int distance, int direction) {
+    public String toString() {
+        String dir = DirectionUtil.intToString(direction);
+        return "Reachable Rolezone at (" + position.x + "/" + position.y + ") with distance " + distance
+                + " in direction " + dir;
+    }
+}
